requirements.txt
pytelegraf
pytradfri[async]
